extend socket-send and socket-receive methods to class USOCKET
EXPERIMENTAL. socket-send and socket-receive methods are only defined
on (specialized on) datagram-usocket.  But the send,sendto and
recv,recvfrom methods are useful even otherwise, say when the socket
is a connected active local unix domain socket.

to do this correctly we could define a `local-socket' class to
represent this or define classes for which send(2) and recv(2) methods
are applicable and derive those from usocket and define these methods
there. but here is the kludge.

* backend/iolib.lisp: define usocket:socket-send and
usocket:socket-receive methods specialized on usocket:socket
* backend/openmcl.lisp: -do-
* backend/sbcl.lisp: -do-

[250609] define ccl::socket-format on fundamental-file-socket-stream
to read from a socket, handles this case

```
socat  UNIX-LISTEN:/tmp/testsock

(setq $a (socket-connect #p"/tmp/testsock" nil :protocol :stream))
(write-line "foo bar" (usocket:socket-stream $a))
(setq $buf (make-array 6 :element-type '(unsigned-byte 8)))
(map 'string 'code-char $buf)
(usocket:socket-receive $a $buf 6)
```

[250609]
* backend/sbcl.lisp: (socket-receive): return an array sized down to
the length of the reply if the no buffer was supplied.

[250609]
* backend/ccl.lisp: (socket-receive): define ccl::socket-format on
fundamental-file-socket-stream to read from a socket, handles this
case