extend socket-send and socket-receive methods to class USOCKET
EXPERIMENTAL. socket-send and socket-receive methods are only defined
on specialized on datagram-usocket.  But they send,sendto and
recv,recvfrom methods are useful even otherwise, say when the socket
is a connected active local unix domain socket.

to do this correctly we could define a `local-socket' class to
represent this or define classes for which send(2) and recv(2) methods
are applicable and derive those from usocket and define these methods
there. but here is the kludge.

* backend/iolib.lisp: define usocket:socket-send and
usocket:socket-receive methods specialized on usocket:socket
* backend/openmcl.lisp: -do-
* backend/sbcl.lisp: -do-