squash! extend socket-send and socket-receive methods to class USOCKET
[250907] handle lispworks

iolib does send recv with with-pointer-to-vector-data, which in
lispworks8 uses hcl:with-pinned-objects.

* backend/iolib: (socket-send): introduce an inefficiency for
lispworks by always consing up an :allocation :pinnable array
for the buffer being sent. (maybe-pinned-array): via this.
(socket-receive): add a comment, stating the burden is on the caller
to send an :allocation :static or :pinnable array in which to receive
the data.